@@ -10,35 +10,35 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.2.9",
-    "@angular/common": "^16.2.9",
-    "@angular/compiler": "^16.2.9",
-    "@angular/core": "^16.2.9",
-    "@angular/forms": "^16.2.9",
-    "@angular/localize": "^16.2.9",
-    "@angular/platform-browser": "^16.2.9",
-    "@angular/platform-browser-dynamic": "^16.2.9",
-    "@angular/router": "^16.2.9",
-    "@fortawesome/angular-fontawesome": "^0.10.1",
-    "@fortawesome/fontawesome-svg-core": "^6.0.0",
-    "@fortawesome/free-brands-svg-icons": "^6.0.0",
-    "@fortawesome/free-regular-svg-icons": "^6.0.0",
-    "@fortawesome/free-solid-svg-icons": "^6.0.0",
-    "@ng-bootstrap/ng-bootstrap": "^15.1.1",
+    "@angular/animations": "^19.2.10",
+    "@angular/common": "^19.2.10",
+    "@angular/compiler": "^19.2.10",
+    "@angular/core": "^19.2.10",
+    "@angular/forms": "^19.2.10",
+    "@angular/localize": "^19.2.10",
+    "@angular/platform-browser": "^19.2.10",
+    "@angular/platform-browser-dynamic": "^19.2.10",
+    "@angular/router": "^19.2.10",
+    "@fortawesome/angular-fontawesome": "^1.0.0",
+    "@fortawesome/fontawesome-svg-core": "^6.7.2",
+    "@fortawesome/free-brands-svg-icons": "^6.7.2",
+    "@fortawesome/free-regular-svg-icons": "^6.7.2",
+    "@fortawesome/free-solid-svg-icons": "^6.7.2",
+    "@ng-bootstrap/ng-bootstrap": "^18.0.0",
     "@popperjs/core": "^2.11.8",
-    "bootstrap": "^5.2.3",
-    "leaflet": "^1.7.1",
+    "bootstrap": "^5.3.6",
+    "leaflet": "^1.9.4",
     "localforage": "^1.10.0",
     "rxjs": "~7.5.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.13.3"
+    "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.2.6",
-    "@angular/cli": "^16.2.6",
-    "@angular/compiler-cli": "^16.2.9",
-    "@types/leaflet": "^1.7.0",
+    "@angular-devkit/build-angular": "^19.2.11",
+    "@angular/cli": "^19.2.11",
+    "@angular/compiler-cli": "^19.2.10",
+    "@types/leaflet": "^1.9.17",
     "@types/node": "^12.11.1",
-    "typescript": "~5.1.6"
+    "typescript": "~5.8.3"
   }
 }
