@@ -10,15 +10,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^19.1.2",
-    "@angular/common": "^19.1.2",
-    "@angular/compiler": "^19.1.2",
-    "@angular/core": "^19.1.2",
-    "@angular/forms": "^19.1.2",
-    "@angular/localize": "^19.1.2",
-    "@angular/platform-browser": "^19.1.2",
-    "@angular/platform-browser-dynamic": "^19.1.2",
-    "@angular/router": "^19.1.2",
+    "@angular/animations": "^19.2.8",
+    "@angular/common": "^19.2.8",
+    "@angular/compiler": "^19.2.8",
+    "@angular/core": "^19.2.8",
+    "@angular/forms": "^19.2.8",
+    "@angular/localize": "^19.2.8",
+    "@angular/platform-browser": "^19.2.8",
+    "@angular/platform-browser-dynamic": "^19.2.8",
+    "@angular/router": "^19.2.8",
     "@fortawesome/angular-fontawesome": "^1.0.0",
     "@fortawesome/fontawesome-svg-core": "^6.7.2",
     "@fortawesome/free-brands-svg-icons": "^6.7.2",
@@ -34,9 +34,9 @@
     "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^19.1.3",
-    "@angular/cli": "^19.1.3",
-    "@angular/compiler-cli": "^19.1.2",
+    "@angular-devkit/build-angular": "^19.2.9",
+    "@angular/cli": "^19.2.9",
+    "@angular/compiler-cli": "^19.2.8",
     "@types/node": "^12.11.1",
     "@types/leaflet": "^1.9.3",
     "jasmine-core": "~4.0.0",
