@@ -1,26 +1,25 @@
 import { HttpClient } from '@angular/common/http';
-import { Injectable } from '@angular/core';
+import { inject, Injectable } from '@angular/core';
 import * as localForage from "localforage";
-import { Observable, from, of } from 'rxjs';
-import { map, switchMap } from 'rxjs/operators';
+import { from, Observable, of } from 'rxjs';
+import { delay, map, switchMap } from 'rxjs/operators';
 import { Playground } from '../model';
 
-localForage.config({ driver: localForage.INDEXEDDB, name: 'state-management', version: 1.0, size: 4980736, storeName: 'state-management-store' });
 
 @Injectable({
   providedIn: 'root'
 })
 export class PlaygroundService {
 
-  constructor(private http: HttpClient) {
-  }
+  #http = inject(HttpClient);
 
   list(): Observable<Playground[]> {
     return from(localForage.getItem<Playground[]>('playgrounds')).pipe(
-      switchMap(playgrounds => playgrounds ? of(playgrounds) : this.http.get<Playground[]>('assets/copenhagen.json')),
+      switchMap(playgrounds => playgrounds ? of(playgrounds) : this.#http.get<Playground[]>('assets/copenhagen.json')),
       switchMap(playgrounds => from(localForage.setItem('playgrounds', playgrounds)).pipe(
         map(() => playgrounds)
       )),
+      delay(3000) // Simulate network delay
     );
   }
 
@@ -41,6 +40,9 @@ export class PlaygroundService {
   }
 
   update(id: string, playground: Partial<Playground>): Observable<Playground> {
+    if (/[13579]$/.test(id)) {
+      throw new Error(`${playground.name} is readonly and cannot be changed`);
+    }
     return this.list().pipe(
       map(playgrounds => playgrounds.map(p => p.id === id ? { ...p, ...playground } : p)),
       switchMap(playgrounds => from(localForage.setItem('playgrounds', playgrounds)).pipe(
@@ -51,6 +53,9 @@ export class PlaygroundService {
   }
 
   delete(id: string): Observable<void> {
+    if (/[13579]$/.test(id)) {
+      throw new Error(`Playground is readonly and cannot be deleted`);
+    }
     return this.list().pipe(
       map(playgrounds => playgrounds.filter(p => p.id !== id)),
       switchMap(playgrounds => from(localForage.setItem('playgrounds', playgrounds)).pipe(
