@@ -1,26 +1,36 @@
-import { AsyncPipe, NgFor, NgIf } from '@angular/common';
-import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output, inject } from '@angular/core';
-import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
-import { Observable } from 'rxjs';
-import { Coordinate, Playground } from '../model';
+import { ChangeDetectionStrategy, Component, computed, inject, input, output, Signal } from '@angular/core';
+import { toSignal } from '@angular/core/rxjs-interop';
+import { FaIconComponent } from '@fortawesome/angular-fontawesome';
+import { Playground } from '../model';
 import { DefaultDescriptionPipe, DistancePipe } from "../pipe";
 import { LocationService } from '../service';
+import { FormControl, ReactiveFormsModule } from '@angular/forms';
 
 @Component({
   selector: 'loop-sidebar',
-  standalone: true,
   templateUrl: './sidebar.component.html',
-  imports: [NgFor, NgIf, AsyncPipe, FontAwesomeModule, DefaultDescriptionPipe, DistancePipe],
+  imports: [FaIconComponent, DefaultDescriptionPipe, DistancePipe, ReactiveFormsModule],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class SidebarComponent {
 
-  @Input() playgrounds: Playground[] | null = [];
-  @Output() selected = new EventEmitter<Playground>();
-  @Output() edit = new EventEmitter<Playground>();
+  playgrounds = input<Playground[] | null>([]);
+  selected = output<Playground>();
+  edit = output<Playground>();
 
   selectedPlayground: Playground | undefined;
-  location$: Observable<Coordinate> = inject(LocationService).location$;
+  location = toSignal(inject(LocationService).location$);
+
+  loading = computed(() => !this.playgrounds()?.length || !this.location());
+  filterControl = new FormControl<string>('');
+  filteredPlaygrounds: Signal<Playground[] | undefined>;
+
+  constructor() {
+    const filter = toSignal(this.filterControl.valueChanges);
+    this.filteredPlaygrounds = computed(() => {
+      return this.playgrounds()?.filter(p => p.name.toLocaleLowerCase().includes(filter()?.toLocaleLowerCase() ?? ''));
+    });
+  }
 
   selectPlayground(playground: Playground): void {
     this.selectedPlayground = playground;
