@@ -10,8 +10,9 @@ const isMarker = (marker: Marker | undefined): marker is Marker => !!marker;
 /* tslint:disable:component-selector-name */
 /* tslint:disable:component-selector-prefix */
 @Component({
-  selector: 'leaflet',
-  template: '<div class="vh-100 vw-100 overflow-hidden" [id]="mapId"></div>',
+    selector: 'leaflet',
+    template: '<div class="vh-100 vw-100 overflow-hidden" [id]="mapId"></div>',
+    standalone: false
 })
 export class LeafletComponent implements AfterViewInit, OnDestroy {
 
